move tree utils to separate file
